requirements.txt
faster-whisper
google-generativeai
cohere
noisereduce
librosa
soundfile
numpy
streamlit
python-dotenv